refactor(frontend): split mobile chat screen responsibilities
